Simple TCP socket plugin (untested).

+++ api/net/sf/briar/api/transport/batch/BatchTransportPlugin.java
@@ -50,10 +50,10 @@ public interface BatchTransportPlugin {
 	boolean shouldPoll();
 
 	/**
-	 * Returns the desired interval in seconds between calls to the plugin's
-	 * poll() method.
+	 * Returns the desired interval in milliseconds between calls to the
+	 * plugin's poll() method.
 	 */
-	int getPollingInterval();
+	long getPollingInterval();
 
 	/**
 	 * Attempts to establish incoming and/or outgoing connections using the
